requirements.txt
python-dotenv==0.17.1
python-telegram-bot==13.5
ptb-firebase-persistence==0.1.2
